Move Register into non-index file
